scripts/phabtalk/requirements.txt
phabricator==0.7.0
lxml==4.4.1
gitpython==3.0.5
